Replace allocation macros with safer functions

+++ src/php/handlers/php_queue_handlers.c
@@ -52,7 +52,7 @@ static HashTable *php_ds_queue_get_gc(zval *obj, zval **gc_data, int *gc_count)
     ds_queue_t *queue = Z_DS_QUEUE_P(obj);
 
     *gc_data  = queue->deque->buffer;
-    *gc_count = (int) queue->deque->capacity;
+    *gc_count = queue->deque->capacity;
 
     return NULL;
 }
